+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,1 @@
-null
+- GitHub Actions Markdown Summary - Thanks to @MattParkerDev !